+++ tests/roots/test-validation-errors/conf.py
@@ -1,0 +1,5 @@
+extensions = ["sphinx_yaq"]
+master_doc = "index"
+project = "YAQ validation errors"
+html_theme = "basic"
+exclude_patterns = []